Fixing clock lines & painting logic
Moved definition of repaintMDROSelect to be with two byte code.
The elipse of MDRCk now follows the midpoint of MARMux.
Created definiton and painting function for MDRECk.
EOMux and its output is now colored yellow.
MemWrite vertical line will now adjust to the displacement of the MemWriteLabel.
MDR{O,E}Muxes are painted green when enabled.
Connected MDROCk, MDRECk to the event loop, so that they can trigger render events.
Moved MDR / MAR registers down another 10 pixels.
The space between MARA and MARB was reduced.
The space between MDRO and MARA was reduced.