Online Installer Updates & Bus fixes
There an now multiple installer config files, to allow each operating system to have its own binary repo.
Created a seperate controller script for the installer.
Makefile now generates appropriate binary repositories for online installer
Makefile now checks to see if installer creation utilities exist, and warns if not.
The data bus arrow in the two byte model is the same width as the address bus, and is now drawn in one function.