ALU Output Fixes
The arrowheads on the ALU output are now correctly colored when the ALU is selected but it has invalid output.
The arrow from ANDZ to Z only is enabled if the ALU has output.
Fixed an unused parameter warning in mainwindow.
Added default values to silence warnings in shapes_one_byte_data_bus.h