Hi-DPI improvements, date & version changes, and installer polish
Disabled 'Add' button on updater tool, as it itsn't need for our purposes.
MSVC Redistributable is now picked based on the version of the compiler used to create the application, and always installed if it is needed.
Added a Hi-DPI fix for windows. This should make the application render correctly regardless of the DPI of the monitor it is displayed on.
Updated version number to 9.1 in aboutpep.cpu, and installer configuration files.
Resized aboutpep form to properly display license.
Cleaned up installer scipts. These still need more comments, but they should be in a much better state now.
Updated copyrigh date across project