Adding new event model classes
Refactored Enu namespace to split up keywords from control signals, and clock lines.
Refactored asm to accomodate these changes.
Microcode object refactored to support unconditional jumps and the new enu split.
Refactored microcde objects to better match the layout of the CPU.
Changed case labels in cpupane to accomodate the enu refactor.
Temporarily disabled object code pane, as it does not work quite yet.
Refactored pep files to use new Enu enums.
Refactored specification to use new Enu enums.

Created the cpustate file, which implements the data and control section of the CPU.

The data section is concerned with mantaining the current values of registers, memory, and control lines. In most ways, this replaces the sim namespace.
It is also able to "clock" the currently enabled control lines given the current state, which replaces most of the functionality of the step...() methods fro mthe cpu pane.

The control section of the cpu is concerned with the micropgoram flow. It handles incrementing the microprogram counter, and preparing the data section's control lines for the next cycle.

One tells the control section to step, debug, or run, and it sets the controls for the data section, and then tells the data section to execute.

Value changes are propogated through notifying signals that one subscribes to, and if one cares about the particular values the were changed, the data section can be asked for the new values.

There is currently no model for the setting the specifications.