Fully Implemented One Byte Bus
VC bits on CPU are most likely still wrong on the CPU at this point, but they should be close.
Fixed handleMainBus code to actually allow reading and writing from memory
Stubbed out unit test postconditions for specification of new model.
Properly set the number of registers in the computer.