Updated connect(...) syntax & style fixes
Replaced old style connect(...) statements with the new QT5 syntax across the project.
Note that not all old connect statements could be replaced, as in some cases we depend on default arguments, which the new method doesn't provide. This is one case where the old syntax is still useful and should still be used.
Attempted to make spacing of arguments across connect(...) statments consistent stylistically.